backend/pyproject.toml
[project]
name = "datalens"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic-settings>=2.7",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.18",
    "pandas>=2.2",
    "openpyxl>=3.1",
    "pyarrow>=18.0",
    "chardet>=5.2",
    "httpx>=0.28",
    "boto3>=1.35",
    "scipy>=1.14",
    "pydantic[email]>=2.12.5",
    "gunicorn>=25.1.0",
]
